Extract subview and some logic to search bar view file/type to be able to reuse it for swift ui
