+++ catowser/catowser/Coordinators/WebContentCoordinator.swift
@@ -54,7 +54,7 @@ final class WebContentCoordinator: Coordinator {
     }
     
     func start() {
-        let webViewController = try? WebViewsEnvironment.shared.reuseManager.controllerFor(site, jsPluginsSource, self, self)
+        let webViewController = try? ViewsEnvironment.shared.reuseManager.controllerFor(site, jsPluginsSource, self, self)
         guard let vc = webViewController else {
             assertionFailure("Failed create new web view for tab")
             return
