+++ catowser/catowser/Extensions/View+Extension.swift
@@ -12,4 +12,17 @@ extension View {
     var isPad: Bool {
         UIDevice.current.userInterfaceIdiom == .pad
     }
+    
+    func resignKeyboardOnDragGesture() -> some View {
+        return modifier(ResignKeyboardOnDragGesture())
+    }
+}
+
+struct ResignKeyboardOnDragGesture: ViewModifier {
+    var gesture = DragGesture().onChanged {_ in
+        UIApplication.shared.endEditing(true)
+    }
+    func body(content: Content) -> some View {
+        content.gesture(gesture)
+    }
 }
