+++ catowser/catowser/Search/SearchBarBaseViewController.swift
@@ -8,91 +8,19 @@
 
 import UIKit
 import CoreBrowser
-import FeaturesFlagsKit
-
-/// The sate of search bar
-enum SearchBarState {
-    /// keyboard and `cancel` button are visible
-    case startSearch
-    /// initial state for new blank tab
-    case blankSearch
-    /// keyboard is hidden and old text is visible
-    case cancelTapped
-    /// when keyboard and all buttons are not displayed
-    case viewMode(title: String, searchAddressContent: String)
-}
 
 protocol SearchBarControllerInterface: AnyObject {
-    /* non optional */ func changeState(to state: SearchBarState, animated: Bool)
-}
-
-fileprivate extension String {
-    static let placeholderText: String = NSLocalizedString("placeholder_searchbar",
-                                                           comment: "when search bar is empty")
+    /* non optional */ func changeState(to state: SearchBarState)
 }
 
 final class SearchBarBaseViewController: BaseViewController {
-
-    /// The search bar view.
-    private let searchBarView: UISearchBar = {
-        let view = UISearchBar(frame: CGRect.zero)
-        ThemeProvider.shared.setup(view)
-        view.placeholder = .placeholderText
-        view.autocapitalizationType = .none
-        view.translatesAutoresizingMaskIntoConstraints = false
-        return view
-    }()
-
-    private let siteNameLabel: UILabel = {
-        let label = UILabel(frame: .zero)
-        label.textColor = .black
-        label.backgroundColor = .white
-        label.textAlignment = .center
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.isUserInteractionEnabled = true
-        return label
-    }()
+    /// main search bar view
+    private let searchBarView: SearchBarLegacyView
     
-    private let dohStateIcon: UILabel = {
-        let label = UILabel(frame: .zero)
-        label.textColor = .black
-        label.font = .italicSystemFont(ofSize: 8)
-        label.backgroundColor = .white
-        label.textAlignment = .center
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.isUserInteractionEnabled = true
-        return label
-    }()
-
-    /// To remember previously entered search query
-    private var searchBarContent: String?
-
-    private lazy var siteNameTapGesture: UITapGestureRecognizer = {
-        // Need to init gesture lazily, if it will  be initialized as a constant
-        // then it will not work :( action is not called.
-        // Problem is with using `self` inside constant,
-        // it seems it is not fully initialized at that point.
-        // https://forums.swift.org/t/self-usage-inside-constant-property/21011
-        // https://stackoverflow.com/questions/50393312/why-can-i-use-self-when-i-initialize-property-with-a-closure
-
-        let tap = UITapGestureRecognizer(target: self, action: .siteNameTap)
-        tap.numberOfTapsRequired = 1
-        tap.numberOfTouchesRequired = 1
-        return tap
-    }()
-
-    private lazy var hiddenLabelConstraint: NSLayoutConstraint = {
-        return siteNameLabel.trailingAnchor.constraint(equalTo: view.leadingAnchor, constant: 0)
-    }()
-
-    private lazy var showedLabelConstraint: NSLayoutConstraint = {
-        return siteNameLabel.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: 0)
-    }()
-    
-    init(_ searchBarDelegate: UISearchBarDelegate) {
-        super.init(nibName: nil, bundle: nil)
-        
+    init(_ searchBarDelegate: UISearchBarDelegate?) {
+        searchBarView = .init(frame: .zero)
         searchBarView.delegate = searchBarDelegate
+        super.init(nibName: nil, bundle: nil)
     }
     
     required init?(coder aDecoder: NSCoder) {
@@ -104,33 +32,11 @@ final class SearchBarBaseViewController: BaseViewController {
     }
 
     override func loadView() {
-        view = UIView()
-
-        view.addSubview(searchBarView)
-        view.addSubview(siteNameLabel)
-        siteNameLabel.addSubview(dohStateIcon)
+        view = searchBarView
     }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        siteNameLabel.addGestureRecognizer(siteNameTapGesture)
-        siteNameLabel.alpha = 0
-
-        searchBarView.topAnchor.constraint(equalTo: view.topAnchor).isActive = true
-        searchBarView.leadingAnchor.constraint(equalTo: view.leadingAnchor).isActive = true
-        searchBarView.trailingAnchor.constraint(equalTo: view.trailingAnchor).isActive = true
-        searchBarView.bottomAnchor.constraint(equalTo: view.bottomAnchor).isActive = true
-
-        siteNameLabel.topAnchor.constraint(equalTo: view.topAnchor).isActive = true
-        siteNameLabel.bottomAnchor.constraint(equalTo: view.bottomAnchor).isActive = true
-        siteNameLabel.widthAnchor.constraint(equalTo: searchBarView.widthAnchor).isActive = true
-        hiddenLabelConstraint.isActive = true
-        
-        dohStateIcon.leadingAnchor.constraint(equalTo: siteNameLabel.leadingAnchor).isActive = true
-        dohStateIcon.topAnchor.constraint(equalTo: siteNameLabel.topAnchor).isActive = true
-        dohStateIcon.bottomAnchor.constraint(equalTo: siteNameLabel.bottomAnchor).isActive = true
-        dohStateIcon.widthAnchor.constraint(equalTo: dohStateIcon.heightAnchor).isActive = true
 
         TabsListManager.shared.attach(self)
     }
@@ -144,7 +50,7 @@ extension SearchBarBaseViewController: TabsObserver {
         // So, assume that `tab` parameter is currently selected
         // and will replace content which is currently disprlayed by search bar
 
-        let state: SearchBarState = .viewMode(title: tab.title, searchAddressContent: tab.searchBarContent)
+        let state: SearchBarState = .viewMode(tab.title, tab.searchBarContent, true)
         changeState(to: state)
     }
 
@@ -153,118 +59,19 @@ extension SearchBarBaseViewController: TabsObserver {
 
         switch content {
         case .site(let site):
-            state = .viewMode(title: site.title, searchAddressContent: site.searchBarContent)
+            state = .viewMode( site.title, site.searchBarContent, false)
         default:
             state = .blankSearch
         }
 
         // run without animation, because label with search query
         // slides when web view has already displayed
-        changeState(to: state, animated: false)
+        changeState(to: state)
     }
 }
 
 extension SearchBarBaseViewController: SearchBarControllerInterface {
-    func changeState(to state: SearchBarState, animated: Bool = true) {
-        switch state {
-        case .startSearch:
-            searchBarView.setShowsCancelButton(true, animated: true)
-            guard searchBarView.text != nil else {
-                break
-            }
-            // need somehow select all text in search bar view
-            prepareForEditMode()
-        case .blankSearch:
-            searchBarContent = nil
-            searchBarView.text = nil
-            siteNameLabel.text = .placeholderText
-            searchBarView.setShowsCancelButton(false, animated: false)
-            searchBarView.resignFirstResponder()
-            // for blank mode it is better to hide label and
-            // make search bar frontmost right away
-            prepareForEditMode()
-        case .cancelTapped:
-            searchBarView.setShowsCancelButton(false, animated: true)
-            searchBarView.resignFirstResponder()
-
-            guard searchBarView.text != nil else {
-                break
-            }
-
-            let dohEnabled = FeatureManager.boolValue(of: .dnsOverHTTPSAvailable)
-            dohStateIcon.text = "\(dohEnabled ? "DoH" : "")"
-            prepareForViewMode(animated: true, animateSecurityView: dohEnabled)
-            // even if search bar now is not visible and
-            // it is under label, need to revert text content in it
-            searchBarView.text = self.searchBarContent
-        case .viewMode(let title, let searchBarContent):
-            searchBarView.setShowsCancelButton(false, animated: animated)
-            searchBarView.text = searchBarContent
-            let dohEnabled = FeatureManager.boolValue(of: .dnsOverHTTPSAvailable)
-            dohStateIcon.text = "\(dohEnabled ? "DoH" : "")"
-            siteNameLabel.text = title
-            prepareForViewMode(animated: animated, animateSecurityView: dohEnabled)
-
-            // remember search query in case if it will be edited
-            self.searchBarContent = searchBarContent
-        }
+    func changeState(to state: SearchBarState) {
+        searchBarView.state = state
     }
-}
-
-private extension SearchBarBaseViewController {
-    @objc func handleSiteNameTap(_ gestureRecognizer: UITapGestureRecognizer) {
-        guard gestureRecognizer.view != nil else { return }
-        guard gestureRecognizer.state == .ended else { return }
-
-        searchBarView.setShowsCancelButton(true, animated: true)
-        searchBarView.becomeFirstResponder()
-        prepareForEditMode()
-    }
-
-    func prepareForViewMode(animated: Bool = true, animateSecurityView: Bool = false) {
-        // Order of disabling/enabling is important
-        // to not to cause errors in layout calculation.
-        // First need to disable and after that enable new one.
-        hiddenLabelConstraint.isActive = false
-        showedLabelConstraint.isActive = true
-        
-        func applyLayout() {
-            siteNameLabel.layoutIfNeeded()
-            searchBarView.alpha = 0
-            siteNameLabel.alpha = 1
-            if animateSecurityView {
-                dohStateIcon.alpha = 1
-            }
-        }
-        
-        if animated {
-            UIView.animate(withDuration: 0.3) {
-                applyLayout()
-            }
-        } else {
-            applyLayout()
-        }
-        
-        searchBarView.resignFirstResponder()
-    }
-
-    func prepareForEditMode(and showKeyboard: Bool = false) {
-        showedLabelConstraint.isActive = false
-        hiddenLabelConstraint.isActive = true
-
-        if showKeyboard {
-            searchBarView.becomeFirstResponder()
-        }
-        siteNameLabel.alpha = 0
-        dohStateIcon.alpha = 0
-        
-        UIView.animate(withDuration: 0.3) {
-            self.siteNameLabel.layoutIfNeeded()
-            self.searchBarView.alpha = 1
-        }
-    }
-}
-
-fileprivate extension Selector {
-    static let siteNameTap = #selector(SearchBarBaseViewController.handleSiteNameTap(_:))
 }
